// appbike-front/appbike-front/src/app/models/uso.interface.ts
// Generated by https://quicktype.io

export interface UsoListResponse {
    content:          Uso[];
    pageable:         Pageable;
    last:             boolean;
    totalPages:       number;
    totalElements:    number;
    first:            boolean;
    size:             number;
    number:           number;
    sort:             Sort;
    numberOfElements: number;
    empty:            boolean;
}

export interface Uso {
    id:          number;
    fechaInicio: string;
    fechaFin:    string;
    coste:       number;
    bicicleta:   string;
    modeloBicicleta: string;
    marcaBicicleta: string;
    estacionFin: string;
    usuario:     string;
}

export interface Pageable {
    pageNumber: number;
    pageSize:   number;
    sort:       Sort;
    offset:     number;
    paged:      boolean;
    unpaged:    boolean;
}

export interface Sort {
    empty:    boolean;
    sorted:   boolean;
    unsorted: boolean;
}


export interface UsoResponse {
    id: number;
    fechaInicio: string;
    fechaFin: null;
    coste: number;
    bicicleta: string;
    estacionFin: string;
    usuario: string;
}

export interface UsoBegin {
    id: number;
    fechaInicio: string;
    bicicleta: string;
    usuario: string;
    estacionInicio: string;
}